v45
Add codec.
Some types not working: ARM, BASE85, LZW.
