+++ analyser/parser.hpp
@@ -1,0 +1,1 @@
+// base for file type detection
