+++ filters/base64filter.hpp
@@ -4,7 +4,7 @@
 class base64Filter: public Filter {
 
 public:
-  base64Filter(std::string n);
+  base64Filter(std::string n, Filetype f=DEFAULT);
   ~base64Filter();
   void encode(File *in, File *out, uint64_t size, uint64_t info);
   uint64_t decode(File *in, File *out,  uint64_t size, uint64_t info);
