+++ filters/cdfilter.hpp
@@ -5,7 +5,7 @@
 class CDFilter: public Filter {
 
 public:
-  CDFilter(std::string n);
+  CDFilter(std::string n, Filetype f=DEFAULT);
   ~CDFilter();
   void encode(File *in, File *out, uint64_t size, uint64_t info);
   uint64_t decode(File *in, File *out,  uint64_t size, uint64_t info);
