+++ filters/defaultfilter.hpp
@@ -2,12 +2,11 @@
 #include "filter.hpp"
 
 
-class MDFFilter: public Filter {
+class DefaultFilter: public Filter {
 
 public:
-  MDFFilter(std::string n, Filetype f=DEFAULT);
-  ~MDFFilter();
+  DefaultFilter(std::string n, Filetype f=DEFAULT);
+  ~DefaultFilter();
   void encode(File *in, File *out, uint64_t size, uint64_t info);
   uint64_t decode(File *in, File *out,  uint64_t size, uint64_t info);
-  
 };
