+++ filters/eolfilter.hpp
@@ -4,7 +4,7 @@
 class EOLFilter: public Filter {
 
 public:
-  EOLFilter(std::string n);
+  EOLFilter(std::string n, Filetype f=DEFAULT);
   ~EOLFilter();
   void encode(File *in, File *out, uint64_t size, uint64_t info);
   uint64_t decode(File *in, File *out,  uint64_t size, uint64_t info);
