+++ filters/img24filter.hpp
@@ -6,7 +6,7 @@
 class Img24Filter: public Filter {
 
 public:
-  Img24Filter(std::string n);
+  Img24Filter(std::string n, Filetype f=DEFAULT);
   ~Img24Filter();
   void encode(File *in, File *out, uint64_t size, uint64_t info);
   uint64_t decode(File *in, File *out,  uint64_t size, uint64_t info);
