+++ filters/rlefilter.hpp
@@ -4,7 +4,7 @@
 class rleFilter: public Filter {
 
 public:
-  rleFilter(std::string n);
+  rleFilter(std::string n, Filetype f=DEFAULT);
   ~rleFilter();
   void encode(File *in, File *out, uint64_t size, uint64_t info);
   uint64_t decode(File *in, File *out,  uint64_t size, uint64_t info);
