+++ filters/zlibfilter.hpp
@@ -8,7 +8,7 @@
 class zlibFilter: public Filter {
 
 public:
-  zlibFilter(std::string n);
+  zlibFilter(std::string n, Filetype f=DEFAULT);
   ~zlibFilter();
   void encode(File *in, File *out, uint64_t size, uint64_t info);
   uint64_t decode(File *in, File *out,  uint64_t size, uint64_t info);
